Move DelayedInputStream into JavaUrl to ensure consistent operation wherever used. Update Url.Connection documentation to specify the expectations around access to the Connection's input and output streams.
