+++ core/src/main/java/com/netflix/msl/io/JavaUrl.java
@@ -15,6 +15,7 @@
  */
 package com.netflix.msl.io;
 
+import java.io.FilterInputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
@@ -27,10 +28,91 @@ import java.net.URLConnection;
  */
 public class JavaUrl implements Url {
     /**
+     * A delayed input stream does not open the real input stream until one of its
+     * its methods is called. This class may be useful in situations where the
+     * connection will not permit use of its output stream after the input stream
+     * is requested.
+     */
+    public class DelayedInputStream extends FilterInputStream {
+        /**
+         * Create a new delayed input stream that will not attempt to
+         * construct the input stream from the URL connection until it is
+         * actually needed (i.e. read from).
+         * 
+         * @param conn backing URL connection.
+         */
+        public DelayedInputStream(final URLConnection conn) {
+            super(null);
+            this.conn = conn;
+        }
+        
+        @Override
+        public int available() throws IOException {
+            if (in == null)
+                in = conn.getInputStream();
+            return super.available();
+        }
+
+        @Override
+        public void close() throws IOException {
+            if (in == null)
+                in = conn.getInputStream();
+            super.close();
+        }
+
+        @Override
+        public synchronized void mark(final int readlimit) {
+        }
+
+        @Override
+        public boolean markSupported() {
+            return false;
+        }
+
+        @Override
+        public int read() throws IOException {
+            if (in == null)
+                in = conn.getInputStream();
+            return in.read();
+        }
+
+        @Override
+        public int read(final byte[] b, final int off, final int len) throws IOException {
+            if (in == null)
+                in = conn.getInputStream();
+            return super.read(b, off, len);
+        }
+
+        @Override
+        public int read(final byte[] b) throws IOException {
+            if (in == null)
+                in = conn.getInputStream();
+            return super.read(b);
+        }
+
+        @Override
+        public synchronized void reset() throws IOException {
+            if (in == null)
+                in = conn.getInputStream();
+            super.reset();
+        }
+
+        @Override
+        public long skip(final long n) throws IOException {
+            if (in == null)
+                in = conn.getInputStream();
+            return super.skip(n);
+        }
+        
+        /** Connection providing the input stream. */
+        private final URLConnection conn;
+    }
+    
+    /**
      * An implementation of the {@link Connection} interface backed by the
      * built-in Java {@link URLConnection} class.
      */
-    public class JavaConnection implements Connection {
+    private class JavaConnection implements Connection {
         /**
          * Create a new Java connection with the backing URL connection.
          * 
@@ -45,7 +127,9 @@ public class JavaUrl implements Url {
          */
         @Override
         public InputStream getInputStream() throws IOException {
-            return conn.getInputStream();
+            // Asking for the URL connection's input stream prevents further
+            // writing to the output stream, so return a delayed input stream.
+            return new DelayedInputStream(conn);
         }
 
         /* (non-Javadoc)
